requirements.txt
uvicorn==0.30.6
python-dotenv==1.0.1
fastapi==0.112.2